Transmition and reception of a NODEACTION: node identifier and position, correctly
Pending direction and speed

+++ GameApplicationLayer/GACommonStructures.swift
@@ -51,7 +51,7 @@ public struct GAPScene {
 }
 
 public struct GAPNodeAction {
-    public var nodeIdentifier   : UInt8 = 0
+    public var nodeIdentifier   : UInt16 = 0
     public var startPoint       = ScenePoint(coordinates: [0,0,0])
     var speed                   : Float32 = 0.0
     var direction               = Vector3D()
